requirements.txt
ultralytics
python-dotenv
inference-cli
inference
inference-sdk
opencv-python
tqdm